@@ -1,5 +1,5 @@
 {
-  "name": "paragonie/paseto",
+  "name": "1o/paseto",
   "description": "Platform-Agnostic Security Tokens",
   "license": "ISC",
   "type": "library",
@@ -20,13 +20,18 @@
     }
   },
   "require": {
-    "php": "^7|^8",
-    "paragonie/constant_time_encoding": "^2",
-    "paragonie/sodium_compat": "^1.7"
+    "ext-gmp": "*",
+    "ext-json": "*",
+    "ext-openssl": "*",
+    "php": "^8.1",
+    "paragonie/constant_time_encoding": ">= 2.6",
+    "paragonie/corner": "^2",
+    "paragonie/easy-ecc": "^1",
+    "paragonie/sodium_compat": "^1.17"
   },
   "require-dev": {
-    "phpunit/phpunit": "^6|^7|^8|^9",
-    "vimeo/psalm": "^3|^4"
+    "phpunit/phpunit": "^9",
+    "vimeo/psalm": "^4"
   },
   "scripts": {
     "full-test": [
@@ -35,5 +40,9 @@
     ],
     "static-analysis": "psalm",
     "test": "phpunit"
+  },
+  "suggest": {
+    "ext-sodium": "Better performance for PASETO v4, secure memory zeroing",
+    "paragonie/paserk": "Key serialization, key wrapping, canonical key identification"
   }
 }
