@@ -7,12 +7,19 @@ use ParagonIE\Paseto\Keys\{
     AsymmetricPublicKey,
     SymmetricKey
 };
-use ParagonIE\Paseto\Keys\Version2\{
-    AsymmetricSecretKey as V2AsymmetricSecretKey,
-    AsymmetricPublicKey as V2AsymmetricPublicKey,
-    SymmetricKey as V2SymmetricKey
+use ParagonIE\Paseto\Keys\Version3\{
+    AsymmetricSecretKey as V3AsymmetricSecretKey,
+    AsymmetricPublicKey as V3AsymmetricPublicKey,
+    SymmetricKey as V3SymmetricKey
 };
+use ParagonIE\Paseto\Keys\Version4\{
+    AsymmetricSecretKey as V4AsymmetricSecretKey,
+    AsymmetricPublicKey as V4AsymmetricPublicKey,
+    SymmetricKey as V4SymmetricKey
+};
+use ParagonIE\Paseto\Exception\ExceptionCode;
 use ParagonIE\Paseto\Exception\InvalidPurposeException;
+use function get_class, hash_equals, in_array;
 
 /**
  * Class Purpose
@@ -21,10 +28,10 @@ use ParagonIE\Paseto\Exception\InvalidPurposeException;
 final class Purpose
 {
     /**
-     * A whitelist of allowed values/modes. This simulates an enum.
+     * An allow-list of allowed values/modes. This simulates an enum.
      * @const array<int, string>
      */
-    const WHITELIST = [
+    const ALLOWLIST = [
         'local',
         'public',
     ];
@@ -45,11 +52,11 @@ final class Purpose
      */
     const SENDING_KEY_MAP = [
         SymmetricKey::class => 'local',
-        V1SymmetricKey::class => 'local',
-        V2SymmetricKey::class => 'local',
+        V3SymmetricKey::class => 'local',
+        V4SymmetricKey::class => 'local',
         AsymmetricSecretKey::class => 'public',
-        V1AsymmetricSecretKey::class => 'public',
-        V2AsymmetricSecretKey::class => 'public'
+        V3AsymmetricSecretKey::class => 'public',
+        V4AsymmetricSecretKey::class => 'public',
     ];
 
     /**
@@ -68,42 +75,48 @@ final class Purpose
      */
     const RECEIVING_KEY_MAP = [
         SymmetricKey::class => 'local',
-        V1SymmetricKey::class => 'local',
-        V2SymmetricKey::class => 'local',
+        V3SymmetricKey::class => 'local',
+        V4SymmetricKey::class => 'local',
         AsymmetricPublicKey::class => 'public',
-        V1AsymmetricPublicKey::class => 'public',
-        V2AsymmetricPublicKey::class => 'public'
+        V3AsymmetricPublicKey::class => 'public',
+        V4AsymmetricPublicKey::class => 'public',
     ];
 
     /**
      * Inverse of EXPECTED_SENDING_KEYS, evaluated and statically cached at
      * runtime.
+     *
      * @var array<string, string>
      */
-    private static $sendingKeyToPurpose;
+    private static array $sendingKeyToPurpose = [];
 
     /**
      * Inverse of EXPECTED_RECEIVING_KEYS, evaluated and statically cached at
      * runtime.
+     *
      * @var array<string, string>
      */
-    private static $receivingKeyToPurpose;
+    private static array $receivingKeyToPurpose = [];
 
     /**
      * @var string
      */
-    private $purpose;
+    private string $purpose;
 
     /**
-     * Allowed values in self::WHITELIST
+     * Allowed values in self::ALLOWLIST
      *
      * @param string $rawString
+     *
      * @throws InvalidPurposeException
      */
     public function __construct(string $rawString)
     {
         if (!self::isValid($rawString)) {
-            throw new InvalidPurposeException('Unknown purpose: ' . $rawString);
+            throw new InvalidPurposeException(
+                'Unknown purpose: ' . $rawString,
+                ExceptionCode::PURPOSE_NOT_LOCAL_OR_PUBLIC
+            );
         }
 
         $this->purpose = $rawString;
@@ -113,7 +126,6 @@ final class Purpose
      * Create a local purpose.
      *
      * @return self
-     * @throws InvalidPurposeException
      */
     public static function local(): self
     {
@@ -124,7 +136,6 @@ final class Purpose
      * Create a public purpose.
      *
      * @return self
-     * @throws InvalidPurposeException
      */
     public static function public(): self
     {
@@ -135,8 +146,8 @@ final class Purpose
      * Given a SendingKey, retrieve the corresponding Purpose.
      *
      * @param SendingKey $key
-     *
      * @return self
+     *
      * @throws InvalidPurposeException
      */
     public static function fromSendingKey(SendingKey $key): self
@@ -146,15 +157,15 @@ final class Purpose
             self::$sendingKeyToPurpose = self::SENDING_KEY_MAP;
         }
 
-        return new self(self::$sendingKeyToPurpose[\get_class($key)]);
+        return new self(self::$sendingKeyToPurpose[get_class($key)]);
     }
 
     /**
      * Given a ReceivingKey, retrieve the corresponding Purpose.
      *
      * @param ReceivingKey $key
-     *
      * @return self
+     *
      * @throws InvalidPurposeException
      */
     public static function fromReceivingKey(ReceivingKey $key): self
@@ -164,7 +175,7 @@ final class Purpose
             self::$sendingKeyToPurpose = self::RECEIVING_KEY_MAP;
         }
 
-        return new self(self::$receivingKeyToPurpose[\get_class($key)]);
+        return new self(self::$receivingKeyToPurpose[get_class($key)]);
     }
 
     /**
@@ -175,7 +186,7 @@ final class Purpose
      */
     public function equals(self $purpose): bool
     {
-        return \hash_equals($purpose->purpose, $this->purpose);
+        return hash_equals($purpose->purpose, $this->purpose);
     }
 
     /**
@@ -187,7 +198,7 @@ final class Purpose
      */
     public static function isValid(string $rawString): bool
     {
-        return \in_array($rawString, self::WHITELIST, true);
+        return in_array($rawString, self::ALLOWLIST, true);
     }
 
     /**
@@ -224,10 +235,7 @@ final class Purpose
      */
     public function expectedSendingKeyType(): string
     {
-        /** @var string */
-        $keyType = self::EXPECTED_SENDING_KEYS[$this->rawString()];
-
-        return $keyType;
+        return self::EXPECTED_SENDING_KEYS[$this->rawString()];
     }
 
     /**
@@ -238,10 +246,7 @@ final class Purpose
      */
     public function expectedReceivingKeyType(): string
     {
-        /** @var string */
-        $keyType = self::EXPECTED_RECEIVING_KEYS[$this->rawString()];
-
-        return $keyType;
+        return self::EXPECTED_RECEIVING_KEYS[$this->rawString()];
     }
 
     /**
