@@ -7,6 +7,7 @@ use ParagonIE\Paseto\{
     ValidationRuleInterface
 };
 use ParagonIE\Paseto\Exception\PasetoException;
+use function hash_equals;
 
 /**
  * Class Subject
@@ -14,11 +15,8 @@ use ParagonIE\Paseto\Exception\PasetoException;
  */
 class Subject implements ValidationRuleInterface
 {
-    /** @var string $failure */
-    protected $failure = 'OK';
-
-    /** @var string $subject */
-    protected $subject;
+    protected string $failure = 'OK';
+    protected string $subject;
 
     /**
      * Subject constructor.
@@ -38,6 +36,8 @@ class Subject implements ValidationRuleInterface
     }
 
     /**
+     * Does the 'sub' claim match what we expect from the Parser?
+     *
      * @param JsonToken $token
      * @return bool
      */
@@ -45,7 +45,7 @@ class Subject implements ValidationRuleInterface
     {
         try {
             $subject = $token->getSubject();
-            if (!\hash_equals($this->subject, $subject)) {
+            if (!hash_equals($this->subject, $subject)) {
                 $this->failure = 'This token was not related to ' .
                     $this->subject . ' (expected); its subject is ' .
                     $subject . ' instead.';
