@@ -7,6 +7,9 @@ use ParagonIE\Paseto\{
     ValidationRuleInterface
 };
 use ParagonIE\Paseto\Exception\PasetoException;
+use Exception;
+use DateTime;
+use DateTimeInterface;
 
 /**
  * Class ValidAt
@@ -14,20 +17,17 @@ use ParagonIE\Paseto\Exception\PasetoException;
  */
 class ValidAt implements ValidationRuleInterface
 {
-    /** @var string $failure */
-    protected $failure = 'OK';
-
-    /** @var \DateTimeInterface $now */
-    protected $now;
+    protected string $failure = 'OK';
+    protected DateTimeInterface $now;
 
     /**
      * ValidAt constructor.
-     * @param \DateTimeInterface|null $now
+     * @param DateTimeInterface|null $now
      */
-    public function __construct(\DateTimeInterface $now = null)
+    public function __construct(DateTimeInterface $now = null)
     {
         if (!$now) {
-            $now = new \DateTime();
+            $now = new DateTime();
         }
         $this->now = $now;
     }
@@ -43,6 +43,8 @@ class ValidAt implements ValidationRuleInterface
     /**
      * @param JsonToken $token
      * @return bool
+     *
+     * @throws Exception
      */
     public function isValid(JsonToken $token): bool
     {
