@@ -17,11 +17,16 @@ interface KeyInterface
     public function getProtocol(): ProtocolInterface;
 
     /**
+     * Returns the raw key as a string.
+     *
      * @return string
      */
-    public function raw();
+    public function raw(): string;
 
     /**
+     * This hides the internal state from var_dump(), etc. if it returns
+     * an empty array.
+     *
      * @return array
      */
     public function __debugInfo();
