@@ -16,6 +16,7 @@ use ParagonIE\Paseto\Keys\Version2\{
     SymmetricKey as V2SymmetricKey
 };
 use ParagonIE\Paseto\Exception\{
+    ExceptionCode,
     InvalidVersionException,
     PasetoException,
     SecurityException
@@ -28,10 +29,24 @@ use ParagonIE\Paseto\Parsing\{
     Header,
     PasetoMessage
 };
+use Exception;
+use SodiumException;
+use Throwable;
+use TypeError;
+use function hash_equals,
+    is_null,
+    is_string,
+    sodium_crypto_generichash,
+    sodium_crypto_aead_xchacha20poly1305_ietf_decrypt,
+    sodium_crypto_aead_xchacha20poly1305_ietf_encrypt,
+    sodium_crypto_sign_detached,
+    sodium_crypto_sign_verify_detached;
 
 /**
- * Class Version2
+ * Class Version1
  * @package ParagonIE\Paseto\Protocol
+ *
+ * @deprecated See Version4 instead.
  */
 class Version2 implements ProtocolInterface
 {
@@ -52,21 +67,35 @@ class Version2 implements ProtocolInterface
      */
     public static function header(): string
     {
-        return self::HEADER;
+        return (string) static::HEADER;
     }
 
     /**
-     * @return int
+     * Does this protocol support implicit assertions?
+     * No.
+     *
+     * @return bool
+     */
+    public static function supportsImplicitAssertions(): bool
+    {
+        return false;
+    }
+
+    /**
+     * @return positive-int
      */
     public static function getSymmetricKeyByteLength(): int
     {
-        return (int) static::SYMMETRIC_KEY_BYTES;
+        return static::SYMMETRIC_KEY_BYTES;
     }
 
     /**
+     * Generate an asymmetric secret key for use with v2.public tokens.
+     *
      * @return AsymmetricSecretKey
-     * @throws \Exception
-     * @throws \TypeError
+     *
+     * @throws Exception
+     * @throws TypeError
      */
     public static function generateAsymmetricSecretKey(): AsymmetricSecretKey
     {
@@ -74,9 +103,12 @@ class Version2 implements ProtocolInterface
     }
 
     /**
+     * Generate a symmetric key for use with v2.local tokens.
+     *
      * @return SymmetricKey
-     * @throws \Exception
-     * @throws \TypeError
+     *
+     * @throws Exception
+     * @throws TypeError
      */
     public static function generateSymmetricKey(): SymmetricKey
     {
@@ -89,15 +121,18 @@ class Version2 implements ProtocolInterface
      * @param string $data
      * @param SymmetricKey $key
      * @param string $footer
+     * @param string $implicit (Unused)
      * @return string
+     *
      * @throws PasetoException
-     * @throws \SodiumException
-     * @throws \TypeError
+     * @throws SodiumException
+     * @throws TypeError
      */
     public static function encrypt(
         string $data,
         SymmetricKey $key,
-        string $footer = ''
+        string $footer = '',
+        string $implicit = ''
     ): string {
         return self::__encrypt($data, $key, $footer);
     }
@@ -108,24 +143,36 @@ class Version2 implements ProtocolInterface
      * @param string $data
      * @param SymmetricKey $key
      * @param string $footer
+     * @param string $implicit (Unused)
      * @param string $nonceForUnitTesting
      * @return string
+     *
      * @throws PasetoException
-     * @throws \SodiumException
-     * @throws \TypeError
+     * @throws SodiumException
+     * @throws TypeError
      */
     protected static function __encrypt(
         string $data,
         SymmetricKey $key,
         string $footer = '',
+        string $implicit = '',
         string $nonceForUnitTesting = ''
     ): string {
+        /*
+         * PASETO Version 2 - Encrypt - Step 1
+         *
+         * We already have a type-safety check on local/public. The following check
+         * constrains the key to v2 tokens only.
+         */
         if (!($key->getProtocol() instanceof Version2)) {
-            throw new InvalidVersionException('The given key is not intended for this version of PASETO.');
+            throw new InvalidVersionException(
+                'The given key is not intended for this version of PASETO.',
+                ExceptionCode::WRONG_KEY_FOR_VERSION
+            );
         }
-        return self::aeadEncrypt(
+        return static::aeadEncrypt(
             $data,
-            self::HEADER . '.local.',
+            static::header() . '.local.', // // PASETO Version 2 - Encrypt - Step 2
             $key,
             $footer,
             $nonceForUnitTesting
@@ -138,21 +185,33 @@ class Version2 implements ProtocolInterface
      * @param string $data
      * @param SymmetricKey $key
      * @param string|null $footer
+     * @param string $implicit (Unused)
      * @return string
      *
      * @throws PasetoException
-     * @throws \SodiumException
-     * @throws \TypeError
+     * @throws SodiumException
+     * @throws TypeError
      */
     public static function decrypt(
         string $data,
         SymmetricKey $key,
-        string $footer = null
+        string $footer = null,
+        string $implicit = ''
     ): string {
+        /*
+         * PASETO Version 2 - Decrypt - Step 1
+         *
+         * We already have a type-safety check on local/public. The following check
+         * constrains the key to v2 tokens only.
+         */
         if (!($key->getProtocol() instanceof Version2)) {
-            throw new InvalidVersionException('The given key is not intended for this version of PASETO.');
+            throw new InvalidVersionException(
+                'The given key is not intended for this version of PASETO.',
+                ExceptionCode::WRONG_KEY_FOR_VERSION
+            );
         }
-        if (\is_null($footer)) {
+        // PASETO Version 2 - Decrypt - Step 2:
+        if (is_null($footer)) {
             $footer = Util::extractFooter($data);
             $data = Util::removeFooter($data);
         } else {
@@ -160,12 +219,15 @@ class Version2 implements ProtocolInterface
         }
         $message = self::aeadDecrypt(
             $data,
-            self::HEADER . '.local.',
+            static::header() . '.local.',
             $key,
-            (string) $footer
+            $footer
         );
-        if (!\is_string($message)) {
-            throw new PasetoException('Invalid message decryption');
+        if (!is_string($message)) {
+            throw new PasetoException(
+                'Invalid message decryption',
+                ExceptionCode::UNSPECIFIED_CRYPTOGRAPHIC_ERROR
+            );
         }
 
         return $message;
@@ -177,25 +239,42 @@ class Version2 implements ProtocolInterface
      * @param string $data
      * @param AsymmetricSecretKey $key
      * @param string $footer
+     * @param string $implicit (Unused)
      * @return string
      *
      * @throws PasetoException
-     * @throws \TypeError
+     * @throws SodiumException
+     * @throws TypeError
      */
     public static function sign(
         string $data,
         AsymmetricSecretKey $key,
-        string $footer = ''
+        string $footer = '',
+        string $implicit = ''
     ): string {
+        /*
+         * PASETO Version 2 - Sign - Step 1
+         *
+         * We already have a type-safety check on local/public. The following check
+         * constrains the key to v2 tokens only.
+         */
         if (!($key->getProtocol() instanceof Version2)) {
-            throw new InvalidVersionException('The given key is not intended for this version of PASETO.');
+            throw new InvalidVersionException(
+                'The given key is not intended for this version of PASETO.',
+                ExceptionCode::WRONG_KEY_FOR_VERSION
+            );
         }
-        $header = self::HEADER . '.public.';
-        $signature = \sodium_crypto_sign_detached(
+
+        // PASETO Version 2 - Sign - Step 2:
+        $header = static::header() . '.public.';
+
+        // PASETO Version 2 - Sign - Step 3, 4:
+        $signature = sodium_crypto_sign_detached(
             Util::preAuthEncode($header, $data, $footer),
             $key->raw()
         );
 
+        // PASETO Version 2 - Sign - Step 5:
         return (new PasetoMessage(
             Header::fromString($header),
             $data . $signature,
@@ -209,33 +288,61 @@ class Version2 implements ProtocolInterface
      * @param string $signMsg
      * @param AsymmetricPublicKey $key
      * @param string|null $footer
+     * @param string $implicit (Unused)
      * @return string
+     *
      * @throws PasetoException
-     * @throws \TypeError
+     * @throws SodiumException
+     * @throws TypeError
      */
     public static function verify(
         string $signMsg,
         AsymmetricPublicKey $key,
-        string $footer = null
+        string $footer = null,
+        string $implicit = ''
     ): string {
+        /*
+         * PASETO Version 2 - Verify - Step 1
+         *
+         * We already have a type-safety check on local/public. The following check
+         * constrains the key to v2 tokens only.
+         */
         if (!($key->getProtocol() instanceof Version2)) {
-            throw new InvalidVersionException('The given key is not intended for this version of PASETO.');
+            throw new InvalidVersionException(
+                'The given key is not intended for this version of PASETO.',
+                ExceptionCode::WRONG_KEY_FOR_VERSION
+            );
         }
-        if (\is_null($footer)) {
+
+        // PASETO Version 2 - Verify - Step 2:
+        if (is_null($footer)) {
             $footer = Util::extractFooter($signMsg);
         } else {
             $signMsg = Util::validateAndRemoveFooter($signMsg, $footer);
         }
         $signMsg = Util::removeFooter($signMsg);
-        /** @var string $footer */
-        $expectHeader = self::HEADER . '.public.';
-        $givenHeader = Binary::safeSubstr($signMsg, 0, 10);
-        if (!\hash_equals($expectHeader, $givenHeader)) {
-            throw new PasetoException('Invalid message header.');
+
+        // PASETO Version 2 - Verify - Step 3:
+        $expectHeader = static::header() . '.public.';
+        $headerLength = Binary::safeStrlen($expectHeader);
+        $givenHeader = Binary::safeSubstr($signMsg, 0, $headerLength);
+        if (!hash_equals($expectHeader, $givenHeader)) {
+            throw new PasetoException(
+                'Invalid message header.',
+                ExceptionCode::INVALID_HEADER
+            );
         }
-        $decoded = Base64UrlSafe::decode(Binary::safeSubstr($signMsg, 10));
+        $decoded = Base64UrlSafe::decode(Binary::safeSubstr($signMsg, $headerLength));
         $len = Binary::safeStrlen($decoded);
 
+        if ($len <= SODIUM_CRYPTO_SIGN_BYTES) {
+            throw new PasetoException(
+                'Invalid message length.',
+                ExceptionCode::INVALID_MESSAGE_LENGTH
+            );
+        }
+
+        // PASETO Version 2 - Verify - Step 4:
         // Separate the decoded bundle into the message and signature.
         $message = Binary::safeSubstr(
             $decoded,
@@ -247,13 +354,19 @@ class Version2 implements ProtocolInterface
             $len - SODIUM_CRYPTO_SIGN_BYTES
         );
 
-        $valid = \sodium_crypto_sign_verify_detached(
+        // PASETO Version 2 - Verify - Step 5, 6:
+        $valid = sodium_crypto_sign_verify_detached(
             $signature,
             Util::preAuthEncode($givenHeader, $message, $footer),
             $key->raw()
         );
+
+        // PASETO Version 2 - Verify - Step 7:
         if (!$valid) {
-            throw new PasetoException('Invalid signature for this message');
+            throw new PasetoException(
+                'Invalid signature for this message',
+                ExceptionCode::INVALID_SIGNATURE
+            );
         }
         return $message;
     }
@@ -269,9 +382,11 @@ class Version2 implements ProtocolInterface
      * @param string $footer
      * @param string $nonceForUnitTesting
      * @return string
+     *
+     * @throws Exception
      * @throws SecurityException
-     * @throws \SodiumException
-     * @throws \TypeError
+     * @throws SodiumException
+     * @throws TypeError
      */
     public static function aeadEncrypt(
         string $plaintext,
@@ -280,25 +395,32 @@ class Version2 implements ProtocolInterface
         string $footer = '',
         string $nonceForUnitTesting = ''
     ): string {
+
+        // PASETO Version 2 - Encrypt - Step 3:
         if ($nonceForUnitTesting) {
             $nonce = $nonceForUnitTesting;
         } else {
-            $nonce = \random_bytes(
-                \SODIUM_CRYPTO_AEAD_XCHACHA20POLY1305_IETF_NPUBBYTES
+            $nonce = random_bytes(
+                SODIUM_CRYPTO_AEAD_XCHACHA20POLY1305_IETF_NPUBBYTES
             );
         }
-        $nonce = \sodium_crypto_generichash(
+
+        // PASETO Version 2 - Encrypt - Step 4:
+        $nonce = sodium_crypto_generichash(
             $plaintext,
             $nonce,
-            \SODIUM_CRYPTO_AEAD_XCHACHA20POLY1305_IETF_NPUBBYTES
+            SODIUM_CRYPTO_AEAD_XCHACHA20POLY1305_IETF_NPUBBYTES
         );
-        $ciphertext = \sodium_crypto_aead_xchacha20poly1305_ietf_encrypt(
+
+        // PASETO Version 2 - Encrypt - Step 5, 6:
+        $ciphertext = sodium_crypto_aead_xchacha20poly1305_ietf_encrypt(
             $plaintext,
             Util::preAuthEncode($header, $nonce, $footer),
             $nonce,
             $key->raw()
         );
 
+        // PASETO Version 2 - Encrypt - Step 7:
         return (new PasetoMessage(
             Header::fromString($header),
             $nonce . $ciphertext,
@@ -313,12 +435,11 @@ class Version2 implements ProtocolInterface
      * @param string $header
      * @param SymmetricKey $key
      * @param string $footer
-     *
      * @return bool|string
      *
      * @throws PasetoException
-     * @throws \SodiumException
-     * @throws \TypeError
+     * @throws SodiumException
+     * @throws TypeError
      */
     public static function aeadDecrypt(
         string $message,
@@ -326,28 +447,47 @@ class Version2 implements ProtocolInterface
         SymmetricKey $key,
         string $footer = ''
     ) {
+        // PASETO Version 2 - Decrypt - Step 3:
         $expectedLen = Binary::safeStrlen($header);
         $givenHeader = Binary::safeSubstr($message, 0, $expectedLen);
-        if (!\hash_equals($header, $givenHeader)) {
-            throw new PasetoException('Invalid message header.');
+        if (!hash_equals($header, $givenHeader)) {
+            throw new PasetoException(
+                'Invalid message header.',
+                ExceptionCode::INVALID_HEADER
+            );
         }
+
+        // PASETO Version 2 - Decrypt - Step 4:
         try {
             $decoded = Base64UrlSafe::decode(Binary::safeSubstr($message, $expectedLen));
-        } catch (\Throwable $ex) {
-            throw new PasetoException('Invalid encoding detected', 0, $ex);
+        } catch (Throwable $ex) {
+            throw new PasetoException(
+                'Invalid encoding detected',
+                ExceptionCode::INVALID_BASE64URL,
+                $ex
+            );
         }
         $len = Binary::safeStrlen($decoded);
+
+        if ($len <= SODIUM_CRYPTO_AEAD_XCHACHA20POLY1305_IETF_NPUBBYTES) {
+            throw new PasetoException(
+                'Invalid message length.',
+                ExceptionCode::INVALID_MESSAGE_LENGTH
+            );
+        }
+
         $nonce = Binary::safeSubstr(
             $decoded,
             0,
-            \SODIUM_CRYPTO_AEAD_XCHACHA20POLY1305_IETF_NPUBBYTES
+            SODIUM_CRYPTO_AEAD_XCHACHA20POLY1305_IETF_NPUBBYTES
         );
         $ciphertext = Binary::safeSubstr(
             $decoded,
-            \SODIUM_CRYPTO_AEAD_XCHACHA20POLY1305_IETF_NPUBBYTES,
-            $len - \SODIUM_CRYPTO_AEAD_XCHACHA20POLY1305_IETF_NPUBBYTES
+            SODIUM_CRYPTO_AEAD_XCHACHA20POLY1305_IETF_NPUBBYTES,
+            $len - SODIUM_CRYPTO_AEAD_XCHACHA20POLY1305_IETF_NPUBBYTES
         );
-        return \sodium_crypto_aead_xchacha20poly1305_ietf_decrypt(
+        // PASETO Version 2 - Decrypt - Step  5, 6, 7:
+        return sodium_crypto_aead_xchacha20poly1305_ietf_decrypt(
             $ciphertext,
             Util::preAuthEncode($header, $nonce, $footer),
             $nonce,
